# Add syntactic filtering of side conditions after simplification