This PR introduces a heuristic in the equation application process:
- after simplifying the predicates in the requires clause, see if an of them is syncatically present in the path condition (which is known to be true) and delete them