@geo2a 2024-05-17
There's a slight slowdown on KEVM with this change.

I've measured with 5 workers and then re-measured with 1 worker to exclude the noise. The results are almost identical, so it looks like https://github.com/runtimeverification/haskell-backend/pull/3853 has indeed worked.

* 5 workers

```
| Test                                    | georgy-second-side-condition-filterin time | master-b27210141 time | (georgy-second-side-condition-filterin/master-b27210141) time
|-----------------------------------------|--------------------------------------------|-----------------------|---------------------------------------------------------------
| mcd/vat-muluu-pass-spec.k               | 95.58                                      | 106.08                | 0.9010180995475113
| mcd/dstoken-approve-fail-rough-spec.k   | 87.8                                       | 92.12                 | 0.9531046461137646
| benchmarks/requires01-a0gt0-spec.k      | 42.37                                      | 44.38                 | 0.9547093285263631
| benchmarks/requires01-a0le0-spec.k      | 42.16                                      | 44.06                 | 0.9568769859282795
| examples/solidity-code-spec.md          | 104.39                                     | 108.08                | 0.965858623242043
| examples/erc20-spec.md                  | 159.87                                     | 165.51                | 0.9659235091535255
| erc20/ds/transferFrom-success-2-spec.k  | 82.37                                      | 79.47                 | 1.0364917578960615
| erc20/hkg/allowance-spec.k              | 44.03                                      | 42.38                 | 1.038933459178858
| erc20/ds/transferFrom-success-1-spec.k  | 87.02                                      | 83.37                 | 1.0437807364759504
| erc20/ds/approve-success-spec.k         | 62.3                                       | 59.19                 | 1.0525426592329785
| erc20/ds/transfer-success-2-spec.k      | 69.46                                      | 63.66                 | 1.0911090166509583
| erc20/ds/transfer-success-1-spec.k      | 72.53                                      | 66.12                 | 1.0969449485783425
| erc20/hkg/transferFrom-success-2-spec.k | 63.49                                      | 55.33                 | 1.1474787637809507
| erc20/hkg/transferFrom-success-1-spec.k | 65.75                                      | 57.16                 | 1.1502799160251924
| erc20/hkg/approve-spec.k                | 50.16                                      | 43.55                 | 1.1517795637198622
| TOTAL                                   | 1129.28                                    | 1110.46               | 1.0169479314878518
```

* 1 worker

```
| Test                                    | georgy-second-side-condition-filterin time | master-b27210141 time | (georgy-second-side-condition-filterin/master-b27210141) time
|-----------------------------------------|--------------------------------------------|-----------------------|---------------------------------------------------------------
| mcd/vat-muluu-pass-spec.k               | 95.12                                      | 106.21                | 0.8955842199416252
| benchmarks/requires01-a0le0-spec.k      | 41.45                                      | 43.94                 | 0.9433318161128813
| benchmarks/requires01-a0gt0-spec.k      | 42.33                                      | 44.59                 | 0.9493159901323166
| mcd/dstoken-approve-fail-rough-spec.k   | 87.17                                      | 91.57                 | 0.9519493283826581
| examples/solidity-code-spec.md          | 103.97                                     | 108.35                | 0.9595754499307799
| examples/erc20-spec.md                  | 159.57                                     | 165.61                | 0.9635287724171244
| erc20/hkg/allowance-spec.k              | 44.09                                      | 42.5                  | 1.0374117647058825
| erc20/ds/transferFrom-success-1-spec.k  | 86.92                                      | 83.52                 | 1.0407088122605366
| erc20/ds/transferFrom-success-2-spec.k  | 82.64                                      | 79.22                 | 1.0431709164352436
| erc20/ds/approve-success-spec.k         | 61.96                                      | 58.62                 | 1.0569771409075401
| erc20/ds/transfer-success-2-spec.k      | 69.52                                      | 64.13                 | 1.0840480274442539
| erc20/ds/transfer-success-1-spec.k      | 72.11                                      | 65.94                 | 1.0935699120412496
| erc20/hkg/approve-spec.k                | 49.68                                      | 43.65                 | 1.1381443298969072
| erc20/hkg/transferFrom-success-1-spec.k | 65.94                                      | 57.27                 | 1.1513881613410162
| erc20/hkg/transferFrom-success-2-spec.k | 63.24                                      | 54.88                 | 1.152332361516035
| TOTAL                                   | 1125.7099999999998                         | 1110.0000000000002    | 1.0141531531531527
```