@geo2a 2024-05-21
I have three recently re-enabled KEVM proofs failing on the feature branch:
```
FAILED src/tests/integration/test_prove.py::test_pyk_prove[mcd/flopper-tick-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_pyk_prove[mcd/flopper-dent-guy-diff-tic-not-0-pass-rough-spec.k]
FAILED src/tests/integration/test_prove.py::test_pyk_prove[mcd/flopper-dent-guy-same-pass-rough-spec.k]
```

Will investigate `mcd/flopper-tick-pass-rough-spec.k`, as it seems to fail relatively quickly (140.20s)