@geo2a 2024-05-22
KEVM performance numbers, excluding `mcd/flopper-tick-pass-rough-spec.k`:

| Test                                                               | georgy-second-side-condition-filterin time | master-0f3148308 time | (georgy-second-side-condition-filterin/master-0f3148308) time
|--------------------------------------------------------------------|--------------------------------------------|-----------------------|---------------------------------------------------------------
| mcd/vat-muluu-pass-spec.k                                          | 105.48                                     | 116.19                | 0.9078233927188226
| mcd/flopper-kick-pass-rough-spec.k                                 | 153.49                                     | 169.05                | 0.9079562259686483
| mcd/end-pack-pass-rough-spec.k                                     | 195.32                                     | 212.04                | 0.921146953405018
| benchmarks/requires01-a0gt0-spec.k                                 | 50.91                                      | 54.0                  | 0.9427777777777777
| examples/solidity-code-spec.md                                     | 122.11                                     | 127.34                | 0.958928851892571
| benchmarks/encode-keccak00-spec.k                                  | 59.6                                       | 62.02                 | 0.9609803289261528
| mcd/end-cash-pass-rough-spec.k                                     | 303.58                                     | 315.39                | 0.9625542978534513
| benchmarks/keccak00-spec.k                                         | 56.98                                      | 59.15                 | 0.9633136094674556
| kontrol/test-expectcalltest-testexpectstaticcall-0-spec.k          | 54.44                                      | 56.39                 | 0.9654194006029437
| benchmarks/storagevar02-nooverflow-spec.k                          | 52.1                                       | 50.17                 | 1.0384692047040063
| erc20/ds/transferFrom-success-2-spec.k                             | 98.47                                      | 94.56                 | 1.041349407783418
| mcd/flopper-cage-pass-spec.k                                       | 62.85                                      | 60.23                 | 1.0434999169848913
| kontrol/test-allowchangestest-testfailallowcallstoaddress-0-spec.k | 58.01                                      | 55.58                 | 1.0437207628643397
| erc20/ds/transferFrom-success-1-spec.k                             | 103.61                                     | 97.09                 | 1.0671541868369554
| erc20/ds/approve-success-spec.k                                    | 75.24                                      | 70.16                 | 1.072405929304447
| mcd/vat-addui-pass-spec.k                                          | 141.23                                     | 131.34                | 1.0753007461550175
| erc20/ds/transfer-success-1-spec.k                                 | 86.57                                      | 79.04                 | 1.0952682186234817
| erc20/ds/transfer-success-2-spec.k                                 | 83.56                                      | 75.88                 | 1.1012124406958357
| mcd/vat-addui-fail-rough-spec.k                                    | 81.57                                      | 69.68                 | 1.170637198622273
| erc20/hkg/approve-spec.k                                           | 61.74                                      | 52.58                 | 1.1742107265119819
| erc20/hkg/transferFrom-success-1-spec.k                            | 79.44                                      | 67.55                 | 1.1760177646188008
| erc20/hkg/transferFrom-success-2-spec.k                            | 75.37                                      | 63.91                 | 1.1793146612423722