@geo2a 2024-05-22
Kontrol performance:
| Test                                                    | georgy-second-side-condition-filterin time | master-0f3148308 time | (georgy-second-side-condition-filterin/master-0f3148308) time
|---------------------------------------------------------|--------------------------------------------|-----------------------|---------------------------------------------------------------
| FreshCheatcodes.test_address()                          | 43.23                                      | 58.98                 | 0.7329603255340793
| AddrTest.test_notBuiltinAddress_symbolic(address)       | 55.83                                      | 65.31                 | 0.8548461185117133
| StructTypeTest.test_vars((uint8,uint32,bytes32))        | 35.72                                      | 40.84                 | 0.8746327130264445
| ConstructorTest.run_constructor()                       | 96.39                                      | 109.73                | 0.8784288708648501
| BlockParamsTest.testFee(uint256)                        | 38.48                                      | 42.76                 | 0.8999064546304958
| BlockParamsTest.testChainId(uint256)                    | 38.68                                      | 42.89                 | 0.9018419211937514
| AccountParamsTest.testDealSymbolic(uint256)             | 43.01                                      | 47.68                 | 0.9020553691275167
| BlockParamsTest.testRoll(uint256)                       | 38.48                                      | 42.54                 | 0.9045604137282557
| ExpectRevertTest.testFail_expectRevert_false()          | 134.45                                     | 145.31                | 0.9252632303351455
| FreshBytesTest.test_symbolic_bytes_length               | 103.04                                     | 109.31                | 0.9426401976031471
| BytesTypeTest.test_bytes4(bytes4)                       | 32.35                                      | 34.22                 | 0.9453535943892462
| BlockParamsTest.testWarp(uint256)-trace_options2        | 30.48                                      | 32.09                 | 0.9498286070426923
| HevmTests.prove_require_assert_false                    | 70.82                                      | 74.2                  | 0.9544474393530996
| CounterTest.testSetNumber(uint256)                      | 145.76                                     | 152.02                | 0.9588212077358241
| AccountParamsTest.testNonceSymbolic(uint64)             | 94.58                                      | 98.61                 | 0.9591319338809451
| Setup2Test.test_setup()                                 | 31.46                                      | 32.8                  | 0.9591463414634147
| FeeTest.test_fee_setup()                                | 65.07                                      | 67.73                 | 0.9607264137014615
| StoreTest.testGasLoadColdVM()                           | 39.03                                      | 40.46                 | 0.9646564508156203
| SetUpTest.testSetupData()                               | 79.38                                      | 76.51                 | 1.0375114364135405
| StoreTest.testLoadNonExistent()                         | 38.73                                      | 36.92                 | 1.0490249187432286
| HevmTests.prove_revert                                  | 23.31                                      | 21.93                 | 1.0629274965800273
| ExpectRevertTest.testFail_ExpectRevert_failAndSuccess() | 115.08                                     | 105.97                | 1.085967726715108
| HevmTests.proveFail_all_branches                        | 71.89                                      | 64.9                  | 1.107704160246533
| TOTAL                                                   | 1465.2499999999998                         | 1543.7100000000003    | 0.9491743915631818

`test_foundry_prove.py::test_foundry_refute_node` fails on both feature and master branches :shrug: