@geo2a 2024-05-30
Measured KEVM performance again: no big difference, long-running proofs seem to benefit though.
| Test                                     | georgy-second-side-condition-filterin time | master-b6ab7a905 time | (georgy-second-side-condition-filterin/master-b6ab7a905) time
|------------------------------------------|--------------------------------------------|-----------------------|---------------------------------------------------------------
| mcd/flopper-tick-pass-rough-spec.k       | 200.67                                     | 225.68                | 0.8891793690180786
| mcd/flopper-kick-pass-rough-spec.k       | 147.07                                     | 163.6                 | 0.898960880195599
| mcd/end-pack-pass-rough-spec.k           | 187.88                                     | 203.43                | 0.9235609300496485
| mcd/vat-muluu-pass-spec.k                | 105.56                                     | 114.24                | 0.9240196078431373
| benchmarks/requires01-a0gt0-spec.k       | 51.15                                      | 55.28                 | 0.9252894356005789
| benchmarks/requires01-a0le0-spec.k       | 51.0                                       | 54.45                 | 0.9366391184573002
| erc20/ds/transferFrom-failure-1-d-spec.k | 55.1                                       | 58.04                 | 0.9493452791178498
| benchmarks/overflow00-nooverflow-spec.k  | 51.71                                      | 54.27                 | 0.9528284503408881
| benchmarks/structarg01-spec.k            | 63.09                                      | 65.92                 | 0.9570691747572816
| examples/erc20-spec.md                   | 179.97                                     | 187.77                | 0.9584598178622783
| mcd/dstoken-approve-fail-rough-spec.k    | 96.62                                      | 100.74                | 0.9591026404605917
| benchmarks/storagevar01-spec.k           | 49.24                                      | 51.15                 | 0.9626588465298144
| bihu/forwardToHotWallet-success-2-spec.k | 60.67                                      | 62.89                 | 0.9647002703132453
| bihu/functional-spec.k                   | 50.7                                       | 48.84                 | 1.038083538083538
| erc20/ds/transferFrom-success-2-spec.k   | 95.69                                      | 90.9                  | 1.0526952695269527
| erc20/ds/transferFrom-success-1-spec.k   | 98.98                                      | 93.41                 | 1.0596295899796597
| erc20/ds/approve-success-spec.k          | 72.96                                      | 68.09                 | 1.0715229842855043
| mcd/vat-addui-pass-spec.k                | 139.59                                     | 129.88                | 1.0747613181398215
| erc20/hkg/approve-spec.k                 | 59.66                                      | 54.34                 | 1.0979020979020977
| erc20/ds/transfer-success-1-spec.k       | 84.08                                      | 75.2                  | 1.1180851063829786
| erc20/ds/transfer-success-2-spec.k       | 81.59                                      | 72.85                 | 1.1199725463280714
| erc20/hkg/transferFrom-success-2-spec.k  | 72.77                                      | 62.23                 | 1.1693716856821468
| erc20/hkg/transferFrom-success-1-spec.k  | 77.54                                      | 66.17                 | 1.1718301345020403
| mcd/vat-addui-fail-rough-spec.k          | 82.32                                      | 69.49                 | 1.1846308821413152
| TOTAL                                    | 2215.6100000000006                         | 2228.86               | 0.9940552569474981